Factor out configure script logic to separate module for clarity
The new module is private for now.
