+++ changelog.d/pr-8648
@@ -1,0 +1,12 @@
+synopsis: Tiny refactor of how Cabal handles configure scripts
+packages: Cabal
+prs: #8648
+
+description: {
+None of this is visible downstream
+
+- Remove needless parameter on one private function.
+
+- Move another internal function (and ones that only it uses from the same module) to new private module.
+
+}
